// src/Models/BankForm/BindEntityDetails.ts
export interface BindEntityDetails {
    IsTodatemandatoryenach: string;
    ISSendEmailCustomer: string;
    IsRefrence2Mandatory: string;
    IsShowMandateMode: string;
    ModeOfPayment: string;
    IsRefrenceNumeric: string;
    amout: string;
    Date: string;
    FromDate: string;
    Name: string;
    UtilityCode: string;
    DebitType: string;
    FrequencyType: string;
    ToDebit: string;
    SponsorBankCode: string;
    PeriodType: string;
 
}