// src/Models/BankForm/UserEntityDetail.ts
export class UserEntityDetail {

    UserId: string;
    EntityId: string;
    AppId: string;
    dataList: any;
    Flag: string;
    FlagValue: string;
    MandateMode: string;
    Bankaccountno: string;
    Catagorycode: string;
    Mandatetype: string;
    UMRN: string;
    UMRNDATE: string;
    Sponsorcode: string;
    Utilitycode: string;
    Authrizename: string;
    Todebit: string; 
    Withbank: string;
    Debittype: string;
    IFSC: string;
    MICR: string;
    Frequency: string;
    Amountrupees: string;
    Amount: string;     
    Email: string;
    Phoneno: string;
    Refrence1: string;
    Refrence2: string;
    PeriodFrom: string;
    PeriodTo: string;
    Untillcancelled: string;
    Customer1: string;
    Customer2: string;
    Customer3: string;
    
}